[Move mocks to appropriate place]
