+++ src/features/ModelPlan/AdditionalQuestionnaires/WaiverAssessmentSurvey/_components/ModelPlanQuestionsForm/index.test.tsx
@@ -8,7 +8,7 @@ import {
   ModelCategory,
   YesNoOtherType
 } from 'gql/generated/graphql';
-import { modelID, modelPlanQuestionsDataMocks } from 'tests/mock/readonly';
+import { modelID, modelPlanQuestionsDataMocks } from 'tests/mock/general';
 
 import ModelPlanQuestionsForm, { ModelPlanQuestionsDataType } from './index';
