+++ src/tests/mock/general.ts
@@ -29,9 +29,15 @@ import {
   GetFavoritesDocument,
   GetFavoritesQuery,
   GetFavoritesQueryVariables,
+  GetMedicarePaymentWaiversDocument,
+  GetMedicarePaymentWaiversQuery,
+  GetMedicarePaymentWaiversQueryVariables,
   GetModelPlanBaseDocument,
   GetModelPlanBaseQuery,
   GetModelPlanBaseQueryVariables,
+  GetModelPlanQuestionsDocument,
+  GetModelPlanQuestionsQuery,
+  GetModelPlanQuestionsQueryVariables,
   GetModelPlansByStatusGroupDocument,
   GetModelPlansByStatusGroupQuery,
   GetModelPlansByStatusGroupQueryVariables,
@@ -45,6 +51,7 @@ import {
   ModelPlanStatusGroup,
   ModelStatus,
   MtoStatus,
+  NotSelectedReason,
   TaskStatus,
   WaiverAssessmentSurveyStatus
 } from 'gql/generated/graphql';
@@ -206,6 +213,101 @@ export const allCommonWaiversMocks: MockedResponse<
         __typename: 'Query',
         commonWaivers: commonWaiversMockData
       }
+    }
+  }
+];
+
+export const modelPlanQuestionsDataMocks: GetModelPlanQuestionsQuery = {
+  __typename: 'Query',
+  modelPlan: {
+    __typename: 'ModelPlan',
+    id: modelID,
+    basics: {
+      id: 'basics-123',
+      __typename: 'PlanBasics',
+      additionalModelCategories: [],
+      cmsCenters: [],
+      cmmiGroups: []
+    },
+    generalCharacteristics: {
+      id: 'char-456',
+      __typename: 'PlanGeneralCharacteristics',
+      existingModel: '12',
+      keyCharacteristics: [],
+      geographiesTargetedTypes: [],
+      geographiesStatesAndTerritories: [],
+      geographiesRegionTypes: [],
+      geographiesTargetedAppliedTo: [],
+      waiversRequiredTypes: []
+    }
+  },
+  modelPlanCollection: [
+    {
+      id: 'uuid-mint-1',
+      modelName: 'B. MINT Model Plan',
+      __typename: 'ModelPlan'
+    }
+  ],
+  existingModelCollection: [
+    {
+      id: 12,
+      modelName: 'A. Existing Model Plan',
+      __typename: 'ExistingModel'
+    }
+  ]
+};
+
+export const modelPlanQuestionsMocks: MockedResponse<
+  GetModelPlanQuestionsQuery,
+  GetModelPlanQuestionsQueryVariables
+>[] = [
+  {
+    request: {
+      query: GetModelPlanQuestionsDocument,
+      variables: { id: modelID }
+    },
+    result: {
+      data: modelPlanQuestionsDataMocks
+    }
+  }
+];
+
+const medicalPaymentWaiversDataMocks: GetMedicarePaymentWaiversQuery = {
+  __typename: 'Query',
+  modelPlan: {
+    __typename: 'ModelPlan',
+    id: modelID,
+    questionnaires: {
+      __typename: 'Questionnaires',
+      waiverAssessmentSurvey: {
+        id: '123',
+        __typename: 'WaiverAssessmentSurvey',
+        modifiesMedicareSavingsPrograms: true,
+        modifiesMedicareSavingsProgramsExample: 'Example',
+        modifiesMedicareSavingsProgramsWhyNot: null,
+        bundlesPayments: false,
+        bundlesPaymentsExample: '',
+        bundlesPaymentsWhyNot: NotSelectedReason.FEEDBACK_AGAINST_USE,
+        offersRiskSharingArrangements: null,
+        offersRiskSharingArrangementsExample: '',
+        offersRiskSharingArrangementsWhyNot: null,
+        suggestedWaivers: []
+      }
+    }
+  }
+};
+
+export const medicalPaymentWaiversMocks: MockedResponse<
+  GetMedicarePaymentWaiversQuery,
+  GetMedicarePaymentWaiversQueryVariables
+>[] = [
+  {
+    request: {
+      query: GetMedicarePaymentWaiversDocument,
+      variables: { id: modelID }
+    },
+    result: {
+      data: medicalPaymentWaiversDataMocks
     }
   }
 ];
